requirements.txt
flask==2.0.1
flask-pymongo==2.3.0
python-dotenv==0.19.0
pymongo==3.12.0
bson==0.5.10
flask-wtf==1.1.1
gunicorn==20.1.0
